Fix the bug of blocked buffer in repo
Because I use select clause in repo to wait on multiple channels and
try to access the value in one of the channels, I cannot use unbuffered
channel. If context.Done() happen first, the unbuffered channel never
get the chance to execute <- receive and this will block the sender of
this channel which is the storage operation in DB!

And then this will block all following same  repo operations!
This take me hours to find out this bug!